[B-Fix] messageã 기능 ìì유저 아이디 부분 수정

+++ RestApi/PartyLog/src/main/java/com/ssafy/partylog/api/service/LetterServiceImpl.java
@@ -22,7 +22,7 @@ public class LetterServiceImpl implements LetterService {
     }
 
     @Override
-    public void addLetter(LetterRequest letterRequest) {
+    public void addLetter(LetterRequest letterRequest, int loginUserNo) {
         LetterEntity letter = new LetterEntity();
 
         String uuid = UUID.randomUUID().toString();
@@ -30,8 +30,7 @@ public class LetterServiceImpl implements LetterService {
         letter.setLetterId(uuid);
         letter.setLetterTitle(letterRequest.getLetterTitle());
         letter.setLetterContent(letterRequest.getLetterContent());
-        //JWT 디코딩 코드 작성
-        letter.setLetterWriter(1);
+        letter.setLetterWriter(loginUserNo);
         letter.setLetterReceiver(letterRequest.getLetterReceiver());
 
         letterRepository.save(letter);
@@ -39,17 +38,18 @@ public class LetterServiceImpl implements LetterService {
 
     @Override
     public void deleteLetter(String letterId) {
-        letterRepository.deleteByLetterId(letterId);
+        if (letterRepository.getLettersByLetterId(letterId).getLetter_writer() == loginUserNo) {
+            letterRepository.deleteByLetterId(letterId);
+        }
     }
 
     @Override
-    public List<LetterResponse> searchLetterList(String type, int year, int offset, int limit) {
-        int userNo = 1;
+    public List<LetterResponse> searchLetterList(String type, int year, int offset, int limit, int loginUserNo) {
         List<LetterResponse> list;
         if(type.equals("writer")) {
-            list = letterRepository.getLettersByWriter(userNo, year, offset, limit);
+            list = letterRepository.getLettersByWriter(loginUserNo, year, offset, limit);
         } else {
-            list = letterRepository.getLettersByReceiver(userNo, year, offset, limit);
+            list = letterRepository.getLettersByReceiver(loginUserNo, year, offset, limit);
         }
         return list;
     }
